move FileDiagnosticTelemetryModule to Base SDK (#1651)
* copy FileDiagnosticsTelemetryModule to BaseSDK

* obsolete old classes

* migrate test classes

* fix for running tests on linux

* typo

* testing linux path

* modifying tests for linux

* changelog
