+++ WEB/Src/HostingStartup/HostingStartup.Net45.Tests/HostingStartup.Net45.Tests.csproj
@@ -1,9 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="15.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildProjectDirectory), 'Test.props'))\Test.props" />
-
   <Import Project="$(PropsRoot)\Test.props" />
-  
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
@@ -57,11 +55,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ClassWithFailingStaticConstructor.cs" />
-    <Compile Include="DiagnosticsTelemetryModuleTest.cs" />
     <Compile Include="GlobalSuppressions.cs" />
-    <Compile Include="InMemoryTraceListener.cs" />
-    <Compile Include="TestEventSource.cs" />
-    <Compile Include="TraceSourceForEventSourceTest.cs" />
     <Compile Include="WebRequestTrackingModuleRegisterTest.cs" />
   </ItemGroup>
   <ItemGroup>
